requirements.txt
pandas
ipython
asyncpg
pyyaml
aiohttp
aiohttp_cors
cython
numpy
scipy
pyasn1
aiodns
lxml
pep8
pylint
google-auth
